requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# LangGraph Orchestration
langgraph>=0.0.26
langchain-core>=0.1.23

# Retrieval & Embeddings
sentence-transformers>=2.3.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2

# LLM Integration
google-generativeai>=0.3.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
pandas>=2.1.0
openpyxl>=3.1.0

# Language & Translation
langdetect>=1.0.9
googletrans==4.0.0rc1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Optional: Web Scraping (commented out for minimal install)
# beautifulsoup4>=4.12.0
# playwright>=1.40.0
